+++ src/components/swiper/SwiperItem.tsx
@@ -1,34 +1,16 @@
-// import { useRef, useEffect, useState } from 'react';
-import { useRef } from 'react';
-
 import Card from '../cards/Card';
 
 type PropsType = {
-  itemIndex:number;
-  onTouchStart(
-    event:React.TouchEvent,
-  ) : void;
-  onTouchEnd:()=> void;
+  itemNumber: number;
 };
 
-export default function SwiperItem(props: PropsType) {
-  const {
-    itemIndex, onTouchStart, onTouchEnd,
-  } = props;
-
-  const itemRef = useRef<HTMLDivElement>(null);
+export default function SwiperItemTest(props:PropsType) {
+  const { itemNumber } = props;
+  console.log(itemNumber);
 
   return (
-    <div
-      className="swiper__item"
-      onTouchStart={(event) => onTouchStart(event)}
-      onTouchEnd={() => onTouchEnd()}
-      role="button"
-      tabIndex={0}
-      ref={itemRef}
-    >
-      {itemIndex}
+    <li className="swiper__item">
       <Card />
-    </div>
+    </li>
   );
 }
