+++ src/styles/components/_burger.scss
@@ -6,6 +6,7 @@
 
   display: block;
   text-align: right;
+  padding: 10px 10px 0 0;
 
   &__btn {
     @include btn-reset;
