+++ src/styles/main.scss
@@ -22,3 +22,5 @@
 @import './components/card/card-animation';
 
 @import './components/swiper';
+
+@import './pages/page-content-wrapper';
