+++ src/styles/utils/_vars.scss
@@ -32,4 +32,5 @@ $page-background: $color-grey_max_light;
 // COMMON VARS
 $border-radius-basic: 10px;
 $medium-screen: 991px;
-$small-screen: 575px;
+$small-screen: 767px;
+$mobile-screen: 575px;
